Switched out the Utils class.
Tried to get rid of MethodMissingWrapperWhiteList too, but that
didn't, y'know, *work*, so it's still here. But! Closer!
